Tried creating a test, but isn't working as expected
