+++ pyproject.toml
@@ -9,6 +9,7 @@ python = "^3.7"
 flask = "^1.1.2"
 python-dotenv = "^0.14.0"
 requests = "^2.25.1"
+pytest = "^6.2.2"
 
 [tool.poetry.dev-dependencies]
