+++ todo_app/app.py
@@ -1,24 +1,12 @@
 from flask import Flask
 from flask import render_template, request, redirect
-from todo_app.data.trello_api import TrelloCard, archive_trello_card, get_trello_cards, get_trello_list_id, move_trello_card, create_trello_card
+from todo_app.data.trello_api import TrelloCard, ViewModel, archive_trello_card, get_trello_cards, get_trello_list_id, move_trello_card, create_trello_card
 from todo_app.flask_config import Config
 import datetime
 
 app = Flask(__name__)
 app.config.from_object(Config)
 
-class ViewModel:
-    def __init__(self, items, trello_list_ids):
-        self._items = items
-        self._trello_list_ids = trello_list_ids
-
-    @property
-    def items(self):
-        return self._items
-
-    @property
-    def trello_list_ids(self):
-        return self._trello_list_ids
 
 @app.route('/')
 def index():
